Fix Fortran ctx type for 64-bit machines
We now always use an integer(8)

Also remove PGA_DEBUG_MAXPGAPACKFUNCTIONS and compute size of debug array
statically. This fixes a bug where the end of the debug array is not
detected.

+++ examples/fortran/classic.f
@@ -18,7 +18,7 @@ c
       integer           NumCoords(3)
       common /data/     Lower, Upper, NumCoords
 
-      integer ctx
+      integer(8)        ctx
       integer           testnum, maxiter, i, ierror
       double precision  l(20), u(20)
 
@@ -63,7 +63,7 @@ c     Yes, it's ugly, but it does work...
 
       double precision function griewank(ctx, p, pop)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, len
       double precision  allele, sum, product
 
@@ -83,7 +83,7 @@ c     Yes, it's ugly, but it does work...
 
       double precision function rastrigin(ctx, p, pop)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, len
       double precision  allele, sum
 
@@ -103,7 +103,7 @@ c     Yes, it's ugly, but it does work...
 
       double precision function schwefel(ctx, p, pop)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, len
       double precision  allele, sum
